refactor: split launcher config and runtime
